[EXPR] ir_operator.h->expr_operator.h Centralize const folder logic
